add v3/memories routes for memories (facts)
