+++ backend/main.py
@@ -5,9 +5,9 @@ import firebase_admin
 from fastapi import FastAPI
 
 from modal import Image, App, asgi_app, Secret
-from routers import workflow, chat, firmware, plugins, memories, transcribe, notifications, \
-    speech_profile, agents, facts, users, processing_memories, trends, sdcard, sync, apps, custom_auth, payment, \
-    integration, conversations
+from routers import workflow, chat, firmware, plugins, memories_deprecated, transcribe, notifications, \
+    speech_profile, agents, facts_deprecated, users, processing_memories, trends, sdcard, sync, apps, custom_auth, \
+    payment, integration, conversations, memories
 
 from utils.other.timeout import TimeoutMiddleware
 
@@ -20,9 +20,10 @@ else:
 
 app = FastAPI()
 app.include_router(transcribe.router)
-app.include_router(memories.router)
+app.include_router(memories_deprecated.router)
 app.include_router(conversations.router)
-app.include_router(facts.router)
+app.include_router(memories.router)
+app.include_router(facts_deprecated.router)
 app.include_router(chat.router)
 app.include_router(plugins.router)
 app.include_router(speech_profile.router)
